requirements.txt
annotated-types==0.6.0
anyio==4.4.0
asgiref==3.8.1
asttokens==2.4.1
backoff==2.2.1
bcrypt==4.2.0
blinker==1.7.0
build==1.2.1
cachetools==5.3.3
certifi==2024.2.2
cffi==1.17.1
charset-normalizer==3.3.2
chroma-hnswlib==0.7.6
chromadb==0.5.15
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
cryptography==43.0.1
debugpy==1.8.7
decorator==5.1.1
Deprecated==1.2.14
dnspython==2.6.1
ecdsa==0.19.0
email_validator==2.2.0
executing==2.1.0
fastapi==0.112.1
filelock==3.13.4
Flask==3.0.3
Flask-Cors==4.0.0
flatbuffers==24.3.25
fpdf==1.7.2
fsspec==2024.3.1
google-ai-generativelanguage==0.6.6
google-api-core==2.18.0
google-api-python-client==2.149.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-generativeai==0.7.2
googleapis-common-protos==1.63.0
grpcio==1.62.1
grpcio-status==1.62.1
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.26.0
humanfriendly==10.0
idna==3.7
importlib_metadata==8.0.0
importlib_resources==6.4.4
ipykernel==6.29.5
ipython==8.28.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jupyter_client==8.6.3
jupyter_core==5.7.2
kubernetes==30.1.0
langchain-chroma==0.1.4
langchain-core==0.3.12
langchain-huggingface==0.1.0
langsmith==0.1.136
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mdurl==0.1.2
mmh3==4.1.0
monotonic==1.6
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.3
numpy==1.26.4
oauthlib==3.2.2
onnxruntime==1.19.0
opentelemetry-api==1.26.0
opentelemetry-exporter-otlp-proto-common==1.26.0
opentelemetry-exporter-otlp-proto-grpc==1.26.0
opentelemetry-instrumentation==0.47b0
opentelemetry-instrumentation-asgi==0.47b0
opentelemetry-instrumentation-fastapi==0.47b0
opentelemetry-proto==1.26.0
opentelemetry-sdk==1.26.0
opentelemetry-semantic-conventions==0.47b0
opentelemetry-util-http==0.47b0
orjson==3.10.7
overrides==7.7.0
packaging==24.0
pandas==2.2.2
parso==0.8.4
passlib==1.7.4
pillow==10.4.0
platformdirs==4.3.6
posthog==3.5.2
prompt_toolkit==3.0.48
proto-plus==1.23.0
protobuf==4.25.3
psutil==6.0.0
psycopg2==2.9.9
pure_eval==0.2.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.7.0
pydantic_core==2.18.1
Pygments==2.18.0
PyMuPDF==1.24.1
PyMuPDFb==1.24.1
pyparsing==3.1.2
PyPika==0.48.9
pypng==0.20220715.0
pyproject_hooks==1.1.0
pyreadline3==3.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
pytz==2024.1
pywin32==308
PyYAML==6.0.1
pyzmq==26.2.0
qrcode==7.4.2
regex==2024.4.16
requests==2.31.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.7.1
rsa==4.9
safetensors==0.4.3
scikit-learn==1.5.1
scipy==1.14.1
sentence-transformers==3.0.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
stack-data==0.6.3
starlette==0.38.2
sympy==1.12
tenacity==9.0.0
threadpoolctl==3.5.0
tokenizers==0.20.1
torch==2.2.2
tornado==6.4.1
tqdm==4.66.2
traitlets==5.14.3
transformers==4.45.2
typer==0.12.4
typing_extensions==4.11.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
uvicorn==0.30.6
watchfiles==0.23.0
wcwidth==0.2.13
websocket-client==1.8.0
websockets==13.0
Werkzeug==3.0.2
wrapt==1.16.0
zipp==3.20.0
